+++ apiexpress/src/controllers/bookControllers.js
@@ -1,3 +1,4 @@
+import NotFound from '../errors/NotFound.js';
 import { author } from '../models/Author.js';
 import book from '../models/Book.js';
 
@@ -11,11 +12,12 @@ class BookController {
         }
     }
     static async listBookById(req, res, next) {
+        
         try {
             const id = req.params.id;
             const fetchedBook = await book.findById(id);
             if (!fetchedBook){
-                res.status(404).json({message : 'book not found'});
+                next(new NotFound('book id not found'));
             }else{
                 res.status(200).json(fetchedBook);
             }
@@ -35,34 +37,49 @@ class BookController {
         }
     }
     static async updateBookById(req, res, next) {
-        try {
-            const id = req.params.id;
-            await book.findByIdAndUpdate(id, req.body);
-            res.status(200).json({ message: `Book Updated: ${book.findById(id)}` });
-        } catch (error) {
-            next(error);
+        const id = req.params.id;
+        const fetchedBook = await book.findById(id);
+        if (fetchedBook){
+            try {
+                await book.findByIdAndUpdate(id, req.body);
+                res.status(200).json({ message: `Book Updated: ${book.findById(id)}` });
+            } catch (error) {
+                next(error);
+            }
+        }else{
+            next(new NotFound('book id not found'));
         }
     }
     static async deleteBookById(req, res, next) {
-        try {
-            const id = req.params.id;
-            await book.findByIdAndDelete(id);
-            res
-                .status(200)
-                .json({ message: `book successfully deleted: ${book.findById(id)}` });
-        } catch (error) {
-            next(error);
+        const id = req.params.id;
+        const fetchedBook = await book.findById(id);
+        if (fetchedBook){
+            try {
+                const id = req.params.id;
+                await book.findByIdAndDelete(id);
+                res
+                    .status(200)
+                    .json({ message: `book successfully deleted: ${book.findById(id)}` });
+            } catch (error) {
+                next(error);
+            }
+        } else {
+            next(NotFound('book not found'));
         }
     }
     static async getBookByPublishingCompany(req, res, next) {
         const publishingCompany = req.query.publishingCompany;
-        try {
-            const bookByPublishingCompany = await book.find({
-                publishingCompany: publishingCompany,
-            });
-            res.status(200).json({ book: bookByPublishingCompany });
-        } catch (error) {
-            next(error);
+        const bookByPublishingCompany = await book.find({
+            publishingCompany: publishingCompany,
+        });
+        if(bookByPublishingCompany){
+            try {
+                res.status(200).json({ book: bookByPublishingCompany });
+            } catch (error) {
+                next(error);
+            }
+        }else{
+            next(new NotFound('book not found'));
         }
     }
 }
